+++ example/android/app/src/main/kotlin/com/mutesasira/flutter_mapbox_navigation_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.mutesasira.flutter_mapbox_navigation_example
+
+import io.flutter.embedding.android.FlutterFragmentActivity
+
+class MainActivity : FlutterFragmentActivity()