+++ example/android/app/src/main/kotlin/com/mutesasira/fluttermapboxnavigationexample/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.mutesasira.fluttermapboxnavigationexample
+
+import io.flutter.embedding.android.FlutterFragmentActivity
+
+class MainActivity: FlutterFragmentActivity() {
+}