+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,6 +5,8 @@
 import FlutterMacOS
 import Foundation
 
+import flutter_mapbox_navigation
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  FlutterMapboxNavigationPlugin.register(with: registry.registrar(forPlugin: "FlutterMapboxNavigationPlugin"))
 }