+++ macos/Classes/FlutterMapboxNavigationPlugin.swift
@@ -1,9 +1,9 @@
-import Cocoa
-import FlutterMacOS
+import Flutter
+import Foundation
 
 public class FlutterMapboxNavigationPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "flutter_mapbox_navigation", binaryMessenger: registrar.messenger)
+    let channel = FlutterMethodChannel(name: "flutter_mapbox_navigation", binaryMessenger: registrar.messenger())
     let instance = FlutterMapboxNavigationPlugin()
     registrar.addMethodCallDelegate(instance, channel: channel)
   }
@@ -12,6 +12,22 @@ public class FlutterMapboxNavigationPlugin: NSObject, FlutterPlugin {
     switch call.method {
     case "getPlatformVersion":
       result("macOS " + ProcessInfo.processInfo.operatingSystemVersionString)
+    case "getDistanceRemaining":
+      result(nil)
+    case "getDurationRemaining":
+      result(nil)
+    case "startFreeDrive":
+      result(false)
+    case "startNavigation":
+      result(false)
+    case "addWayPoints":
+      result(false)
+    case "finishNavigation":
+      result(false)
+    case "enableOfflineRouting":
+      result(false)
+    case "registerRouteEventListener":
+      result(nil)
     default:
       result(FlutterMethodNotImplemented)
     }